# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "30Shine Booking Web API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get list of bookings",
                "description": "Returns list of bookings",
                "operationId": "getBookingsList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Booking",
                "description": "Returns status, mess, booking_id",
                "operationId": "postBookingsStore",
                "parameters": [
                    {
                        "name": "stylist_id",
                        "in": "path",
                        "description": "Stylist ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "store_id",
                        "in": "path",
                        "description": "Store ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "promotion_id",
                        "in": "path",
                        "description": "Promotion ID, Default 1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2 or null"
                    },
                    {
                        "name": "service_id_list[]",
                        "in": "path",
                        "description": "List Service ID",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        },
                        "example": "1,2,3,4,5"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Booking Detail",
                "description": "Returns detail booking",
                "operationId": "showDetailBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/bookings/{id}/cancel": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "cancel Booking",
                "description": "",
                "operationId": "cancelBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Booking ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    }
}